+++ array/zfp/pointer3.h
@@ -52,9 +52,9 @@ protected:
   ptrdiff_t offset(ptrdiff_t d = 0) const { return static_cast<ptrdiff_t>(x - container->min_x() + container->size_x() * (y - container->min_y() + container->size_y() * (z - container->min_z()))) + d; }
   void index(size_t& x, size_t& y, size_t& z, ptrdiff_t p) const
   {
-    x = container->min_x() + p % container->size_x(); p /= container->size_x();
-    y = container->min_y() + p % container->size_y(); p /= container->size_y();
-    z = container->min_z() + p;
+    x = container->min_x() + size_t(p % ptrdiff_t(container->size_x())); p /= container->size_x();
+    y = container->min_y() + size_t(p % ptrdiff_t(container->size_y())); p /= container->size_y();
+    z = container->min_z() + size_t(p);
   }
   void advance(ptrdiff_t d) { index(x, y, z, offset(d)); }
   void increment()
